+++ custom_components/ipx800/switch.py
@@ -77,7 +77,7 @@ class VirtualOutSwitch(IpxDevice, SwitchEntity):
 
     @property
     def is_on(self) -> bool:
-        return self.coordinator.data[f"R{self._id}"] == 1
+        return self.coordinator.data[f"VO{self._id}"] == 1
 
     def turn_on(self, **kwargs):
         self.control.on()