+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.8.2
+
+- Fix polling key for VirtualOut switch
+
 ## 1.8.1
 
 - Remove the wrong climate setup error